% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGraphMulti.R
\name{findGraphMulti}
\alias{findGraphMulti}
\title{Causal discovery with high dimensional non-Gaussian data}
\usage{
findGraphMulti(Y, maxInDegree = 3, degree = 3, B = NULL,
  cutOffScaling = 0.5, verbose = T)
}
\arguments{
\item{Y}{n x p matrix of observations with sample in row, variable in column}

\item{maxInDegree}{Assumed largest in-degree}

\item{degree}{The degree of the moment which is non-Gaussianity (i.e., this is K in the paper)}

\item{B}{true adjacency matrix if available (used primarily for debugging and simulations)}

\item{cutOffScaling}{alpha value for pruning away false parents}

\item{verbose:}{Print progress updates}
}
\value{
\item{topOrder}{Estimated topological ordering of variables}
   \item{prune}{final values of tau statistics}
   \item{cutoff}{final value of cut off parameter}
   \item{parents}{estimated parent sets}
}
\description{
Estimates a causal ordering for high dimensional SEM with non-Gaussian errors
}
