% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDAG_HUB.R
\name{rDAG_hub}
\alias{rDAG_hub}
\title{Causal discovery with high dimensional non-Gaussian data}
\usage{
rDAG_hub(p, n, lowScale = 0.8, highScale = 1, lowEdge = 0.65,
  highEdge = 1, nHubs = 3, hubEdge = 0.2)
}
\arguments{
\item{p}{the number of variables in the graph}

\item{n}{the number of samples of the p variables}

\item{lowScale}{the standard deviation for the error term for each node is drawn uniformly from [lowScale, highScale]}

\item{highScale}{the standard deviation for the error term for each node is drawn uniformly from [lowScale, highScale]}

\item{lowEdge}{the coefficient for each edge v -> v+1 is drawn uniformly from +/-[lowEdge, highEdge]}

\item{highEdge}{the coefficient for each edge v -> v+1 is drawn uniformly from +/-[lowEdge, highEdge]}

\item{nHubs}{the number of hub nodes}

\item{hubEdge}{the coefficient for each edge from a hub to a downstream node is set to hubEdge. This should be appropriately small
so the variance of Y does not blow up}
}
\value{
\item{B}{the random coefficient matrix}
   \item{Y.unif}{the sampled data}
   \item{scale.param}{the error standard deviations}
}
\description{
Generates data corresponding to a Hub graph as described in Wang and Drton (2019)
}
